Added License Statement;
Formatted the initial contents of the README file by applying proper spaces and newlines for better readability and aesthetics. 
Added the License Statement for easy access to the LICENSE file from README directly.
As of this commit, this repository uses GNU GPL V3.0  Strong CopyLeft Licensing. (more about this in my docs that I will be creating soon after).