Created README.md;
Created a basic README.md that has the basic HTML code for showcasing and linking the GitHub Badges to respective data.
Additional changes in the file will be done as an when the files that needs to be uploaded are in their right place.
Will be trying some graphically coded GIFs kind of thing as "README" for its font! (commented code at the start of the file of this commit is a step towards that)
OR
Would like to at least settle in having the cool pattern font made with equal to, underscores, and glyphs (etc.)? 
For Instance: 
======
       |
____||